🧹 Organize datastructures
